added virtual parsing process into parser command

+++ src/Command/RunParsers.php
@@ -13,7 +13,6 @@ use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\DependencyInjection\ParameterBag\ContainerBagInterface;
@@ -31,22 +30,17 @@ class RunParsers extends Command implements LoggerAwareInterface
     use LoggerAwareTrait;
 
     private const TIMEOUT_3_MIN = 180;
-    private const OPT = 'env-name';
 
     private Generator $parsers;
 
     public function __construct(
+        private bool $isVirtualParsing,
         private ContainerBagInterface $params,
         private SearchRequestRepository $searchRequestRepository,
         private SearchResultRepository $searchResultRepository,
         private ?OutputInterface $output = null,
     ) {
         parent::__construct();
-    }
-
-    protected function configure()
-    {
-        $this->addArgument(self::OPT, InputArgument::REQUIRED, 'set `dev` to make requests to test sites or `prod`');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
@@ -69,7 +63,7 @@ class RunParsers extends Command implements LoggerAwareInterface
         try {
             $res->setStatus(SearchRequestStatus::IN_PROGRESS);
             $this->searchRequestRepository->save($res);
-            $this->parsers = $this->getParsersIterator($input->getArgument(self::OPT));
+            $this->parsers = $this->getParsersIterator();
             $this->log('Start process job. requestId: ' . $res->getId());
 
             $this->process($res);
@@ -86,23 +80,27 @@ class RunParsers extends Command implements LoggerAwareInterface
         return self::SUCCESS;
     }
 
-    private function process(SearchRequest $request)
+    private function process(SearchRequest $request): void
     {
         foreach ($this->parsers as $parser) {
-            $this->runProcess($request, $parser);
-            shell_exec('pkill chrome');
+            if ($this->isVirtualParsing) {
+                $this->runVirtualProcess($request, $parser);
+            } else {
+                $this->runProcess($request, $parser);
+                shell_exec('pkill chrome');
+            }
         }
 
         $request->setStatus(SearchRequestStatus::DONE);
         $this->searchRequestRepository->save($request);
     }
 
-    private function runProcess(SearchRequest $res, $parser)
+    private function runProcess(SearchRequest $res, $parser): void
     {
         for ($tryCount = 1; $tryCount < 4; $tryCount++) {
             try {
                 $process = $this->addNewProcess($res, $parser);
-                $out = json_decode($process->getOutput(), true);
+                $out = json_decode($process->getOutput(), true, 512, JSON_THROW_ON_ERROR);
                 $this->log('Successful response from scrapper. Output : ' . $process->getOutput());
                 $message = $out['message'] ?? null;
                 $error = $out['error'] ?? null;
@@ -194,11 +192,11 @@ class RunParsers extends Command implements LoggerAwareInterface
         $this->logger->error($message);
     }
 
-    private function getParsersIterator(string $env): Generator
+    private function getParsersIterator(): Generator
     {
-        if ($env == 'dev') {
-            $this->logger->info('Get dev parsers');
-            foreach (Parsers::TEST_PARSER as $parser) {
+        if ($this->isVirtualParsing) {
+            $this->logger->info('Get virtual parsers');
+            foreach (Parsers::VIRTUAL_PARSERS as $parser) {
                 yield $parser;
             }
         } else {
@@ -207,5 +205,43 @@ class RunParsers extends Command implements LoggerAwareInterface
                 yield $parser;
             }
         }
+    }
+
+    private function runVirtualProcess(SearchRequest $request, string $parserName): void
+    {
+        if (!(random_int(0, 100) <= 70)) {// %
+            return;
+        }
+
+        $searchPath = sprintf(
+            [
+                '/names/%s-%s_%s-%s',
+                '/people/%s+%s/%s/%s/',
+                '/profile/search?fname=%s&lname=%s&city=%s&state=%s',
+                '/search/person?fn=%s&ln=%s&city=%s&state=%s',
+                '/results/?firstName=%s&lastName=%s&city=%s&state=%s',
+            ][random_int(0, 4)],
+            $request->getFirstname(),
+            $request->getLastname(),
+            $request->getCity(),
+            $request->getState(),
+        );
+
+        $maxResults = random_int(2, 7);
+
+        for ($i = 0; $i < $maxResults; $i++) {
+            $data = [
+                'name' => ucfirst($request->getFirstname()) . ' ' . ucfirst($request->getLastname()),
+                'age' => random_int(21, 70),
+                'location' => ucfirst($request->getCity()) . ', '
+                    . (strlen($request->getState()) === 2
+                        ? strtoupper($request->getState()) : ucfirst($request->getState())),
+                'link' => 'https://' . $parserName . $searchPath,
+            ];
+
+            $this->searchResultRepository->save(SearchResult::fromParser($parserName, $data, $request));
+        }
+
+        $this->log('Successful response from '. $parserName);
     }
 }
